requirements.txt
asgiref==3.5.0
colorama==0.4.4
colorlog==6.6.0
colormap==1.0.4
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.11
django-crispy-forms==1.14.0
django-import-export==2.7.1
djangorestframework==3.13.1
easydev==0.12.0
et-xmlfile==1.1.0
fonttools==4.29.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib==3.5.1
numpy==1.22.2
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pandas==1.4.0
pexpect==4.8.0
Pillow==9.0.1
ptyprocess==0.7.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
reports==0.4.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
tablib==3.2.0
threadpoolctl==3.1.0
xlrd==2.0.1
xlwt==1.3.0
